pyproject.toml
[project]
name = "doc-scrape"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "agno>=1.5.0",
    "aiofiles>=24.1.0",
    "aiohttp>=3.11.16",
    "asyncio>=3.4.3",
    "beautifulsoup4>=4.13.3",
    "crawl4ai>=0.5.0.post8",
    "httpx>=0.28.1",
    "lxml>=5.3.1",
    "openai>=1.61.1",
    "pgvector>=0.4.1",
    "psycopg[binary]>=3.2.9",
    "python-dotenv>=1.0.1",
    "spider-client>=0.1.32",
    "sqlalchemy>=2.0.40",
]

[dependency-groups]
dev = [
    "ruff>=0.11.2",
]
